requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
itsdangerous
flask-limiter
serverless-wsgi
Werkzeugd
flask-mail